Move matching function to the analyser
