+++ src/ActionCreator.js
@@ -9,34 +9,8 @@ let makeDeleteAction = (snap) => {
 let makeCreationActions = (volume, snapList) => {
 	// volume = attachSnapsToVolume(volume, snapList);
 
-	console.log(`(${volume.VolumeId}): '${volume.Name}' has ${volume.Snapshots} related snapshots`);
 	return determineBackupsNeeded(volume, snapList).map(backup => backup);
 };
-
-// Given a list of EBS volumes and list of snapshots, attaches the snapshot to the
-// EBS's Snapshots[BackupType] array. Returns an object containing the volume list
-// and a list of orphaned snapshots
-var matchSnapsToVolumes = (volumes, snapList) => {
-
-	let matchedVolumes = volumes.map((volume) => {
-		volume.Snapshots = {}
-		snapList = snapList.filter((snap) => {
-			// console.log(snap);
-			if (snap.FromVolumeName === volume.Name) {
-				if (!volume.Snapshots[snap.BackupType]) volume.Snapshots[snap.BackupType] = [];
-				volume.Snapshots[snap.BackupType].push(snap);
-				return false;
-			} else {
-				return true;
-			}
-		});
-		return volume;
-
-	});
-
-	let orphanedSnaps = snapList;
-	return {matchedVolumes, orphanedSnaps};
-}
 
 // Read the tags on the EBS volume and check if the backup requirements are
 // fulfilled by the current state of snapshots. For example: if the volume
@@ -46,4 +20,4 @@ let determineBackupsNeeded = (volume, snapList) => {
 	return [{volume, snapList}];
 };
 
-export {makeDeleteAction, matchSnapsToVolumes, makeCreationActions, determineBackupsNeeded};
+export {makeDeleteAction, makeCreationActions, determineBackupsNeeded};
